6/30

+++ FirstProject/src/login/TCalendar.java
@@ -9,6 +9,8 @@ import java.awt.Font;
 import java.awt.GridLayout;
 import java.awt.event.ActionEvent;
 import java.awt.event.ActionListener;
+import java.awt.event.ItemEvent;
+import java.awt.event.ItemListener;
 import java.awt.event.WindowAdapter;
 import java.awt.event.WindowEvent;
 import java.util.ArrayList;
@@ -37,7 +39,6 @@ class TCalendar extends JFrame implements ActionListener {
 	int year, month, day, todays, memoday = 0;
 
 	Font f;
-	Color bc, fc;
 	Calendar today, cal;
 
 	JButton btnBefore, btnAfter, btnAdd, btnDel, badd, bdel;
@@ -48,7 +49,7 @@ class TCalendar extends JFrame implements ActionListener {
 	JPanel panWest, panSouth, panEast, panNorth;
 
 	JTextField txtMonth, txtYear, txtTime;
-	JTextArea txtschool,txtschedule;
+	JTextArea txtschool, txtschedule;
 	BorderLayout bLayout = new BorderLayout();
 
 	public TCalendar() {
@@ -56,8 +57,9 @@ class TCalendar extends JFrame implements ActionListener {
 		cal = new GregorianCalendar();
 		year = today.get(Calendar.YEAR);
 		month = today.get(Calendar.MONTH) + 1;// 1월의 값이 0
-		
+
 		panNorth = new JPanel();
+		panNorth.setBackground(Color.white);
 		panNorth.add(btnBefore = new JButton("Before"));
 		panNorth.add(txtYear = new JTextField(year + "년"));
 		panNorth.add(txtMonth = new JTextField(month + "월"));
@@ -65,7 +67,7 @@ class TCalendar extends JFrame implements ActionListener {
 		txtMonth.setEnabled(false);
 
 		panNorth.add(btnAfter = new JButton("After"));
-		f = new Font("맑은고딕", Font.BOLD, 26);
+		f = new Font("NEXON Lv1 Gothic", Font.BOLD, 26);
 		txtYear.setFont(f);
 		txtMonth.setFont(f);
 		panNorth.add(btnAdd = new JButton("일정추가"));
@@ -73,9 +75,17 @@ class TCalendar extends JFrame implements ActionListener {
 
 		add(panNorth, "North");
 
+		// 오늘 날짜 출력
+		JLabel lb1 = new JLabel();
+		lb1.setFont(new Font("NEXON Lv1 Gothic", Font.BOLD, 15));
+		String t1 = "오늘날짜 : " + today.get(Calendar.YEAR) + " 년 " + (today.get(Calendar.MONTH) + 1) + " 월 "
+				+ today.get(Calendar.DAY_OF_MONTH) + " 일";
+		lb1.setText(t1);
+		lb1.setForeground(Color.orange);
+		panNorth.add(lb1);
+
 		// 달력
 		panWest = new JPanel(new GridLayout(7, 7));// 격자나,눈금형태의 배치관리자
-		f = new Font("맑은고딕", Font.BOLD, 15);
 
 		gridInit();
 		calSet();
@@ -83,44 +93,54 @@ class TCalendar extends JFrame implements ActionListener {
 		add(panWest, "West");
 
 		panEast = new JPanel();
+		panEast.setBackground(Color.white);
 		panEast.setLayout(new FlowLayout());
+		add(panEast);
+
+		f = new Font("NEXON Lv1 Gothic", Font.BOLD, 13);
+
 		panEast.add(jlSchool = new JLabel("학교명"));
-//		jlSchool.setBounds(300, 200, 50, 50);
+		jlSchool.setFont(f);
+
 		txtschool = new JTextArea();
-//		txtschool.setBounds(300, 200, 140, 350);
-		txtschool.setPreferredSize(new Dimension(140, 280));
-		panEast.add(txtschool);
-		
-		JScrollPane scroll = new JScrollPane(txtschool);
-		panEast.add(scroll);
-//		getContentPane().panEast.add(scroll);
-		
-//		txtschool.add(scroll);
-		
-		
-//		add(panEast, "Center");
-		getContentPane().add(panEast, "Center");
-		
+		txtschool.setPreferredSize(new Dimension(130, 800));
 		txtschool.setEditable(false);
 		txtschool.setFocusable(false);
-		txtschool.setFont(new Font("맑은고딕",1,15));
+		txtschool.setFont(new Font("NEXON Lv1 Gothic", Font.BOLD, 13));
+		panEast.add(txtschool);
+
+		JScrollPane nameScroll = new JScrollPane(txtschool); // JTextArea를 JScrollPane로 감싸기
+		nameScroll.setPreferredSize(new Dimension(180, 300)); // 스크롤 패널의 우선 크기 설정
+		nameScroll.setVerticalScrollBarPolicy(JScrollPane.VERTICAL_SCROLLBAR_AS_NEEDED);
+		panEast.add(nameScroll); // JScrollPane를 패널에 추가
 
 		panEast.add(jlScedule = new JLabel("오늘일정"));
-		panEast.add(txtschedule = new JTextArea());
-		txtschedule.setPreferredSize(new Dimension(140, 280));
-		add(panEast, "Center");
+		jlScedule.setFont(f);
+
+		txtschedule = new JTextArea();
+		txtschedule.setPreferredSize(new Dimension(130, 800));
 		txtschedule.setEditable(false);
 		txtschedule.setFocusable(false);
-		txtschedule.setFont(new Font("맑은고딕",1,15));
+		txtschedule.setFont(new Font("NEXON Lv1 Gothic", Font.BOLD, 13));
+		panEast.add(txtschedule);
+
+		JScrollPane scheduleScroll = new JScrollPane(txtschedule);
+		scheduleScroll.setPreferredSize(new Dimension(180, 300));
+		scheduleScroll.setVerticalScrollBarPolicy(JScrollPane.VERTICAL_SCROLLBAR_AS_NEEDED);
+		panEast.add(scheduleScroll);
 
 		btnBefore.addActionListener(this);
+		btnBefore.setFont(f);
 		btnAfter.addActionListener(this);
+		btnAfter.setFont(f);
 		btnAdd.addActionListener(this);
+		btnAdd.setFont(f);
 		btnDel.addActionListener(this);
+		btnDel.setFont(f);
 
 		setDefaultCloseOperation(JFrame.EXIT_ON_CLOSE);
 		setTitle("시흥시 일정관리 시스템 - 선생님");
-		setBounds(630, 380, 800, 400);
+		setBounds(590, 380, 850, 400);
 		setVisible(true);
 	}
 
@@ -129,7 +149,6 @@ class TCalendar extends JFrame implements ActionListener {
 		cal.set(Calendar.MONTH, (month - 1));
 		cal.set(Calendar.DATE, 1);
 		int dayOfWeek = cal.get(Calendar.DAY_OF_WEEK);
-
 		/*
 		 * get 및 set 를 위한 필드치로, 요일을 나타냅니다. 이 필드의 값은
 		 * SUNDAY,MONDAY,TUESDAY,WEDNESDAY,THURSDAY,FRIDAY, 및 SATURDAY get()메소드의 의해 요일이
@@ -165,11 +184,8 @@ class TCalendar extends JFrame implements ActionListener {
 					calBtn[i + 6 + hopping].setForeground(new Color(0, 0, 255));
 				}
 			}
-
-			/*
-			 * 요일을 찍은 다음부터 계산해야 하니 요일을 찍은 버튼의 갯수를 더하고 인덱스가 0부터 시작이니 -1을 해준 값으로 연산을 해주고 버튼의
-			 * 색깔을 변경해준다.
-			 */
+			// 요일을 찍은 다음부터 계산해야 하니 요일을 찍은 버튼의 갯수를 더하고 인덱스가 0부터 시작이니 -1을 해준 값으로 연산을 해주고 버튼의
+			// 색깔을 변경해준다.
 			calBtn[i + 6 + hopping].setText((i) + "");
 		}
 	}
@@ -203,6 +219,8 @@ class TCalendar extends JFrame implements ActionListener {
 			af.setLayout(null);
 			af.setVisible(true);
 			af.setSize(500, 300);
+			af.getContentPane().setBackground(Color.white);
+			af.setFont(f);
 			af.addWindowListener(new WindowAdapter() {
 				public void windowClosing(WindowEvent e) {
 					af.dispose();
@@ -210,9 +228,16 @@ class TCalendar extends JFrame implements ActionListener {
 			});
 			af.setLocationRelativeTo(null);
 
+			JLabel l = new JLabel("※ 날짜를 선택하지 않을 시, 오늘 날짜가 입력됩니다.");
+			l.setBounds(100, 30, 350, 20);
+			l.setFont(new Font("NEXON Lv1 Gothic", Font.PLAIN, 12));
+			l.setForeground(Color.red);
+			af.add(l);
+
 			UtilDateModel model = new UtilDateModel();
 			JDatePanelImpl datePanel = new JDatePanelImpl(model);
 			JDatePickerImpl datePicker = new JDatePickerImpl(datePanel);
+			datePicker.setBackground(Color.white);
 			af.add(datePicker);
 			datePicker.setBounds(100, 50, 280, 40);
 			datePicker.setVisible(true);
@@ -221,39 +246,69 @@ class TCalendar extends JFrame implements ActionListener {
 			cd.add("초등학교");
 			cd.add("중학교");
 			cd.add("고등학교");
-			cd.setBounds(40, 100, 100, 40);
+			cd.setBounds(35, 100, 100, 40);
+			cd.setFont(f);
 			af.add(cd);
 
 			MainDAO md = new MainDAO();
 			ArrayList<MainVo> chosc = md.chosc(getName());
-			System.out.println(chosc.size());
 			csh = new Choice();
-			for (int i = 0; i < chosc.size(); i++) {
-				csh.add(chosc.get(i).getName());
-			}
-			csh.setBounds(170, 100, 100, 40);
+			csh.add("구분을 선택해주세요.");
+			cd.addItemListener(new ItemListener() {
+				public void itemStateChanged(ItemEvent e) {
+					if (e.getStateChange() == ItemEvent.SELECTED) {
+						String idx = cd.getSelectedItem().toString();
+						if (idx.equals("초등학교")) {
+							csh.removeAll();
+							for (int i = 0; i < 50; i++) {
+								csh.add(chosc.get(i).getName());
+							}
+						} else if (idx.equals("중학교")) {
+							csh.removeAll();
+							for (int i = 51; i < 76; i++) {
+								csh.add(chosc.get(i).getName());
+							}
+						} else if (idx.equals("고등학교")) {
+							csh.removeAll();
+							for (int i = 76; i < 94; i++) {
+								csh.add(chosc.get(i).getName());
+							}
+						}
+					}
+				}
+			});
+
+			csh.setBounds(165, 100, 160, 40);
+			csh.setFont(f);
 			af.add(csh);
 
 			csc = new Choice();
 			csc.add("졸업식");
 			csc.add("공사");
 			csc.add("체험학습");
-			csc.setBounds(320, 100, 90, 40);
+			csc.setBounds(355, 100, 90, 40);
+			csc.setFont(f);
 			af.add(csc);
 
 			badd = new JButton("일정추가");
 			badd.setBounds(180, 180, 100, 30);
+			badd.setBackground(Color.white);
+			badd.setFont(f);
 			af.add(badd);
 			badd.addActionListener(new ActionListener() {
 				public void actionPerformed(ActionEvent e) {
-					JOptionPane.showMessageDialog(null, "일정이 추가되었습니다.", "PLAIN_MESSAGE", JOptionPane.PLAIN_MESSAGE);
-					String div = cd.getSelectedItem();
-					String name = csh.getSelectedItem();
-					String schedule = csc.getSelectedItem();
-					int year = model.getYear();
-					int month = model.getMonth() +1;
-					int day = model.getDay();
-					dao.insert(div, name, schedule,year,month,day);
+					if (csh.getSelectedItem().toString().equals("구분을 선택해주세요.")) {
+						JOptionPane.showMessageDialog(null, "구분을 선택해주세요.", "ERROR", JOptionPane.ERROR_MESSAGE);
+					} else {
+						String div = cd.getSelectedItem();
+						String name = csh.getSelectedItem();
+						String schedule = csc.getSelectedItem();
+						int year = model.getYear();
+						int month = model.getMonth() + 1;
+						int day = model.getDay();
+						dao.insert(div, name, schedule, year, month, day);
+						JOptionPane.showMessageDialog(null, "일정이 추가되었습니다.", "PLAIN_MESSAGE", JOptionPane.PLAIN_MESSAGE);
+					}
 				}
 			});
 
@@ -267,6 +322,8 @@ class TCalendar extends JFrame implements ActionListener {
 			df.setLayout(null);
 			df.setVisible(true);
 			df.setSize(500, 300);
+			df.getContentPane().setBackground(Color.white);
+			df.setFont(f);
 			df.addWindowListener(new WindowAdapter() {
 				public void windowClosing(WindowEvent e) {
 					df.dispose();
@@ -275,72 +332,114 @@ class TCalendar extends JFrame implements ActionListener {
 
 			df.setLocationRelativeTo(null);
 
+			JLabel l = new JLabel("※ 날짜를 선택하지 않을 시, 오늘 날짜가 입력됩니다.");
+			l.setBounds(100, 30, 350, 20);
+			l.setFont(new Font("NEXON Lv1 Gothic", Font.PLAIN, 12));
+			l.setForeground(Color.red);
+			df.add(l);
+
 			UtilDateModel model = new UtilDateModel();
 			JDatePanelImpl datePanel = new JDatePanelImpl(model);
 			JDatePickerImpl datePicker = new JDatePickerImpl(datePanel);
 			df.add(datePicker);
 			datePicker.setBounds(100, 50, 280, 40);
+			datePicker.setBackground(Color.white);
 			datePicker.setVisible(true);
 
 			cd = new Choice();
 			cd.add("초등학교");
 			cd.add("중학교");
 			cd.add("고등학교");
-			cd.setBounds(40, 100, 100, 40);
+			cd.setBounds(35, 100, 100, 40);
+			cd.setFont(f);
 			df.add(cd);
 
 			MainDAO md = new MainDAO();
 			ArrayList<MainVo> chosc = md.chosc(getName());
 			csh = new Choice();
-			for (int i = 0; i < chosc.size(); i++) {
-				csh.add(chosc.get(i).getName());
-			}
-			csh.setBounds(170, 100, 100, 40);
+			csh.add("구분을 선택해주세요.");
+			cd.addItemListener(new ItemListener() {
+				public void itemStateChanged(ItemEvent e) {
+					if (e.getStateChange() == ItemEvent.SELECTED) {
+						String idx = cd.getSelectedItem().toString();
+						if (idx.equals("초등학교")) {
+							csh.removeAll();
+							for (int i = 0; i < 50; i++) {
+								csh.add(chosc.get(i).getName());
+							}
+						} else if (idx.equals("중학교")) {
+							csh.removeAll();
+							for (int i = 50; i < 76; i++) {
+								csh.add(chosc.get(i).getName());
+							}
+						} else if (idx.equals("고등학교")) {
+							csh.removeAll();
+							for (int i = 76; i < 94; i++) {
+								csh.add(chosc.get(i).getName());
+							}
+						}
+					}
+				}
+			});
+
+			csh.setBounds(165, 100, 160, 40);
+			csh.setFont(f);
 			df.add(csh);
-				
+
 			csc = new Choice();
 			csc.add("졸업식");
 			csc.add("공사");
 			csc.add("체험학습");
-			csc.setBounds(320, 100, 90, 40);
+			csc.setBounds(355, 100, 90, 40);
+			csc.setFont(f);
 			df.add(csc);
 
 			bdel = new JButton("일정삭제");
 			bdel.setBounds(180, 180, 100, 30);
+			bdel.setBackground(Color.white);
+			bdel.setFont(f);
 			df.add(bdel);
+
 			bdel.addActionListener(new ActionListener() {
 				public void actionPerformed(ActionEvent e) {
-					JOptionPane.showMessageDialog(null, "일정이 삭제되었습니다.", "PLAIN_MESSAGE", JOptionPane.PLAIN_MESSAGE);
 					String div = cd.getSelectedItem();
 					String name = csh.getSelectedItem();
 					String schedule = csc.getSelectedItem();
 					int year = model.getYear();
-					int month = model.getMonth() +1;
+					int month = model.getMonth() + 1;
 					int day = model.getDay();
-					dao.delete(div,name,schedule,year,month,day);
+					if (csh.getSelectedItem().toString().equals("구분을 선택해주세요.")) {
+						JOptionPane.showMessageDialog(null, "구분을 선택해주세요.", "ERROR", JOptionPane.ERROR_MESSAGE);
+					} else if (dao.delete(div, name, schedule, year, month, day) == 0) {
+						JOptionPane.showMessageDialog(null, "해당 일정은 없습니다.", "ERROR", JOptionPane.ERROR_MESSAGE);
+					} else {
+						dao.delete(div, name, schedule, year, month, day);
+						JOptionPane.showMessageDialog(null, "일정이 삭제되었습니다.", "PLAIN_MESSAGE", JOptionPane.PLAIN_MESSAGE);
+					}
 				}
 			});
-
 		} else if (Integer.parseInt(ae.getActionCommand()) >= 1 && Integer.parseInt(ae.getActionCommand()) <= 31) {
 			day = Integer.parseInt(ae.getActionCommand());
 			// 버튼의 밸류 즉 1,2,3.... 문자를 정수형으로 변환하여 클릭한 날짜를 바꿔준다.
+			// 선택한 버튼 색깔 변경 -> 다른거 누르면 다시 돌아가야하는데 안됨 ㅋㅋ ㅠ
+//			calBtn[day+10].setBackground(new Color(153, 204, 255));
 			System.out.println(day);
 			calSet();
 
 			// 달력에 일자 누르면 DB에 저장된 일정뜸
 			ArrayList<MainVo> listca = dao.listca(year, month, day);
 
-			if (listca.size()!=0) {
-				for(int i=0;i<listca.size();i++) {
+			if (listca.size() != 0) {
+				for (int i = 0; i < listca.size(); i++) {
 					MainVo data = (MainVo) listca.get(i);
 					String sd = data.getSch();
 					String nd = data.getName();
-					if(i==0) {
-						txtschool.setText((i+1)+" "+ nd+"\n");
-						txtschedule.setText((i+1) +" "+ sd+"\n");
+					if (i == 0) {
+						txtschool.setText((i + 1) + " " + nd + "\n");
+						txtschedule.setText((i + 1) + " " + sd + "\n");
 					} else {
-						txtschool.append((i+1)+" "+nd+"\n");
-						txtschedule.append((i+1)+" "+sd+"\n");
+						txtschool.append((i + 1) + " " + nd + "\n");
+						txtschedule.append((i + 1) + " " + sd + "\n");
 					}
 				}
 			} else {
@@ -349,12 +448,13 @@ class TCalendar extends JFrame implements ActionListener {
 			}
 		}
 	}
-	
+
 	public void hideInit() {
 		for (int i = 0; i < calBtn.length; i++) {
 			if ((calBtn[i].getText()).equals(""))
 				calBtn[i].setEnabled(false);
 			// 일이 찍히지 않은 나머지 버튼을 비활성화 시킨다.
+			calBtn[i].setBackground(Color.white);
 		}
 	}
